# internal/server/v1/user/server.gen.go
// Package user provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package user

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	strictgin "github.com/oapi-codegen/runtime/strictmiddleware/gin"
)

const (
	CookieAuthScopes = "CookieAuth.Scopes"
)

// Defines values for DeleteAccountParamsConfirm.
const (
	True DeleteAccountParamsConfirm = true
)

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// UserDataExport Complete export of user data for GDPR compliance
type UserDataExport struct {
	Checklists []struct {
		CreatedAt *time.Time `json:"createdAt,omitempty"`
		Id        *uint      `json:"id,omitempty"`
		Items     *[]struct {
			Completed   *bool   `json:"completed,omitempty"`
			Id          *uint   `json:"id,omitempty"`
			Name        *string `json:"name,omitempty"`
			OrderNumber *int    `json:"orderNumber,omitempty"`
			Rows        *[]struct {
				Completed *bool   `json:"completed,omitempty"`
				Id        *uint   `json:"id,omitempty"`
				Name      *string `json:"name,omitempty"`
			} `json:"rows,omitempty"`
		} `json:"items,omitempty"`
		Name *string `json:"name,omitempty"`

		// Shares Users with whom this checklist is shared
		Shares *[]struct {
			PermissionLevel  *string    `json:"permissionLevel,omitempty"`
			SharedAt         *time.Time `json:"sharedAt,omitempty"`
			SharedWithUserId *string    `json:"sharedWithUserId,omitempty"`
		} `json:"shares,omitempty"`
	} `json:"checklists"`

	// ExportedAt Timestamp of data export
	ExportedAt time.Time `json:"exportedAt"`

	// UserId Google User ID (anonymous identifier)
	UserId string `json:"userId"`
}

// XClientId defines model for X-Client-Id.
type XClientId = string

// DeleteAccountParams defines parameters for DeleteAccount.
type DeleteAccountParams struct {
	// Confirm Must be set to 'true' to confirm deletion
	Confirm DeleteAccountParamsConfirm `form:"confirm" json:"confirm"`
}

// DeleteAccountParamsConfirm defines parameters for DeleteAccount.
type DeleteAccountParamsConfirm bool

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Delete user account (GDPR Right to Erasure)
	// (DELETE /api/v1/user/account)
	DeleteAccount(c *gin.Context, params DeleteAccountParams)
	// Export user data (GDPR Data Portability)
	// (GET /api/v1/user/data-export)
	ExportUserData(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// DeleteAccount operation middleware
func (siw *ServerInterfaceWrapper) DeleteAccount(c *gin.Context) {

	var err error

	c.Set(CookieAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params DeleteAccountParams

	// ------------- Required query parameter "confirm" -------------

	if paramValue := c.Query("confirm"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument confirm is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "confirm", c.Request.URL.Query(), &params.Confirm)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter confirm: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteAccount(c, params)
}

// ExportUserData operation middleware
func (siw *ServerInterfaceWrapper) ExportUserData(c *gin.Context) {

	c.Set(CookieAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ExportUserData(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.DELETE(options.BaseURL+"/api/v1/user/account", wrapper.DeleteAccount)
	router.GET(options.BaseURL+"/api/v1/user/data-export", wrapper.ExportUserData)
}

type DeleteAccountRequestObject struct {
	Params DeleteAccountParams
}

type DeleteAccountResponseObject interface {
	VisitDeleteAccountResponse(w http.ResponseWriter) error
}

type DeleteAccount200JSONResponse struct {
	Message *string `json:"message,omitempty"`
}

func (response DeleteAccount200JSONResponse) VisitDeleteAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type DeleteAccount400JSONResponse Error

func (response DeleteAccount400JSONResponse) VisitDeleteAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type DeleteAccount401JSONResponse Error

func (response DeleteAccount401JSONResponse) VisitDeleteAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type DeleteAccount500JSONResponse Error

func (response DeleteAccount500JSONResponse) VisitDeleteAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type ExportUserDataRequestObject struct {
}

type ExportUserDataResponseObject interface {
	VisitExportUserDataResponse(w http.ResponseWriter) error
}

type ExportUserData200JSONResponse UserDataExport

func (response ExportUserData200JSONResponse) VisitExportUserDataResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ExportUserData401JSONResponse Error

func (response ExportUserData401JSONResponse) VisitExportUserDataResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type ExportUserData500JSONResponse Error

func (response ExportUserData500JSONResponse) VisitExportUserDataResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Delete user account (GDPR Right to Erasure)
	// (DELETE /api/v1/user/account)
	DeleteAccount(ctx context.Context, request DeleteAccountRequestObject) (DeleteAccountResponseObject, error)
	// Export user data (GDPR Data Portability)
	// (GET /api/v1/user/data-export)
	ExportUserData(ctx context.Context, request ExportUserDataRequestObject) (ExportUserDataResponseObject, error)
}

type StrictHandlerFunc = strictgin.StrictGinHandlerFunc
type StrictMiddlewareFunc = strictgin.StrictGinMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// DeleteAccount operation middleware
func (sh *strictHandler) DeleteAccount(ctx *gin.Context, params DeleteAccountParams) {
	var request DeleteAccountRequestObject

	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteAccount(ctx, request.(DeleteAccountRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteAccount")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(DeleteAccountResponseObject); ok {
		if err := validResponse.VisitDeleteAccountResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// ExportUserData operation middleware
func (sh *strictHandler) ExportUserData(ctx *gin.Context) {
	var request ExportUserDataRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.ExportUserData(ctx, request.(ExportUserDataRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ExportUserData")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(ExportUserDataResponseObject); ok {
		if err := validResponse.VisitExportUserDataResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}