# openapi/api_v1.yaml
openapi: 3.0.3
info:
  title: Checklist service v1
  description: Checklist service v1
  version: 1.0.0
security:
  - CookieAuth: []
paths:
...
  /v1/events/checklist-item-updates/{checklistId}:
    get:
      summary: Server-Sent Events stream for real-time updates for checklist items, filtered by checklistId
      operationId: getEventsStreamForChecklistItems
      parameters:
        - name: checklistId
          in: path
          required: true
          schema:
            type: number
            x-go-type: uint
            minimum: 1
            format: int64
          description: Checklist ID to filter events
        - name: clientId
          in: query
          required: false
          schema:
            type: string
          description: Client identifier passed by frontend
      tags:
        - events
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/EventEnvelope'
...
  /api/v1/user/account:
    delete:
      summary: Delete user account (GDPR Right to Erasure)
      operationId: deleteAccount
      description: Permanently delete your account and all associated data (checklists, items, shares). This action is irreversible and implements GDPR Article 17 (Right to Erasure).
      tags:
        - user
      parameters:
        - name: confirm
          in: query
          required: true
          schema:
            type: boolean
            enum: [true]
          description: Must be set to 'true' to confirm deletion
      responses:
        '200':
          description: Account successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Account deleted successfully
        '400':
          description: Missing or invalid confirmation parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/user/data-export:
    get:
      summary: Export user data (GDPR Data Portability)
      operationId: exportUserData
      description: Download all your data in machine-readable JSON format. Implements GDPR Article 20 (Right to Data Portability).
      tags:
        - user
      responses:
        '200':
          description: User data export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDataExport'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: user_token
      description: |
        Cookie-based authentication using Google ID tokens.
        The user_token cookie is automatically sent by the browser and contains a JWT token.
        This cookie is set with HttpOnly, Secure, and SameSite attributes for security.
...
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          nullable: false
          minLength: 0
      required:
        - code
        - message
...
    ChecklistItemResponse:
      type: object
      properties:
        id:
          type: number
          x-go-type: uint
          nullable: false
          minimum: 1
          format: int64
        name:
          type: string
          nullable: false
          minLength: 1
        completed:
          type: boolean
          nullable: false
          default: false
        orderNumber:
          type: number
          x-go-type: uint
          nullable: false
          minimum: 1
          format: int64
        rows:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItemRowResponse'
      required:
        - name
        - completed
        - id
        - orderNumber
        - rows
    ChecklistItemRowResponse:
      type: object
      properties:
        id:
          type: number
          format: int64
          minimum: 0
          x-go-type: uint
          nullable: false
        name:
          type: string
          nullable: false
          minLength: 1
        completed:
          type: boolean
          nullable: true
          default: false
      required:
        - id
        - name
        - completed
...
    EventEnvelope:
      type: object
      description: |
        Envelope for SSE events; sent as JSON in the SSE data field.
        The `type` field indicates the event type, and the `payload` field contains the event data.
        The expected structure of `payload` for each `type` is as follows:
          - checklistItemCreated: ChecklistItemResponse
          - checklistItemUpdated: ChecklistItemResponse
          - checklistItemDeleted: ChecklistItemDeletedEventPayload
          - checklistItemRowAdded: ChecklistItemRowResponse
          - checklistItemRowUpdated: ChecklistItemRowResponse
          - checklistItemRowDeleted: ChecklistItemRowDeletedEventPayload
          - checklistItemRowAdded: ChecklistItemRowAddedEventPayload
          - checklistItemReordered: ChecklistItemReorderedEventPayload
        For event types not listed above, `payload` may be null or a free-form object.
      properties:
        type:
          type: string
          description: Event type identifier
          example: checklistItemCreated
          enum:
            - checklistItemCreated
            - checklistItemUpdated
            - checklistItemDeleted
            - checklistItemRowAdded
            - checklistItemRowUpdated
            - checklistItemRowDeleted
            - checklistItemReordered
        payload:
          description: |
            Payload structure depends on event type:
              - checklistItemCreated, checklistItemUpdated: ChecklistItemResponse
              - checklistItemDeleted: ChecklistItemDeletedEventPayload
              - checklistItemRowAdded, checklistItemRowUpdated: ChecklistItemRowResponse
              - checklistItemRowDeleted: ChecklistItemRowDeletedEventPayload
              - checklistItemReordered: ChecklistItemReorderedEventPayload
          anyOf:
            - $ref: '#/components/schemas/ChecklistItemResponse'
            - $ref: '#/components/schemas/ChecklistItemRowResponse'
            - $ref: '#/components/schemas/ChecklistItemRowDeletedEventPayload'
            - $ref: '#/components/schemas/ChecklistItemRowAddedEventPayload'
            - $ref: '#/components/schemas/ChecklistItemDeletedEventPayload'
            - $ref: '#/components/schemas/ChecklistItemReorderedEventPayload'
      required:
        - type
    ChecklistItemRowDeletedEventPayload:
      type: object
      properties:
        itemId:
          type: number
          x-go-type: uint
          nullable: false
          format: int64
          minimum: 1
        rowId:
          type: number
          x-go-type: uint
          nullable: false
          format: int64
          minimum: 1
      required:
        - itemId
        - rowId
    ChecklistItemRowAddedEventPayload:
      type: object
      properties:
        itemId:
          type: number
          x-go-type: uint
          nullable: false
          format: int64
          minimum: 1
        row:
          $ref: '#/components/schemas/ChecklistItemRowResponse'
      required:
        - itemId
        - row
    ChecklistItemDeletedEventPayload:
      type: object
      properties:
        itemId:
          type: number
          x-go-type: uint
          nullable: false
          format: int64
          minimum: 1
      required:
        - itemId
    ChecklistItemReorderedEventPayload:
      type: object
      properties:
        itemId:
          type: number
          x-go-type: uint
          nullable: false
          format: int64
          minimum: 1
        newOrderNumber:
          type: number
          x-go-type: uint
          nullable: false
          format: int64
          minimum: 1
        orderChanged:
          type: boolean
          description: Indicates if the order number was changed
      required:
        - itemId
        - orderChanged
        - newOrderNumber
...
    UserDataExport:
      type: object
      description: Complete export of user data for GDPR compliance
      properties:
        userId:
          type: string
          description: Google User ID (anonymous identifier)
        exportedAt:
          type: string
          format: date-time
          description: Timestamp of data export
        checklists:
          type: array
          items:
            type: object
            properties:
              id:
                type: number
                x-go-type: uint
                format: int64
              name:
                type: string
              createdAt:
                type: string
                format: date-time
              items:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      x-go-type: uint
                      format: int64
                    name:
                      type: string
                    completed:
                      type: boolean
                    orderNumber:
                      type: integer
                    rows:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: number
                            x-go-type: uint
                            format: int64
                          name:
                            type: string
                          completed:
                            type: boolean
              shares:
                type: array
                description: Users with whom this checklist is shared
                items:
                  type: object
                  properties:
                    sharedWithUserId:
                      type: string
                    permissionLevel:
                      type: string
                    sharedAt:
                      type: string
                      format: date-time
      required:
        - userId
        - exportedAt
        - checklists
...